REFACTOR: Split code into multiple modules ADD: Ability to show left border to highlight more
